requirements.txt
flask
python-dotenv
psycopg2
pymongo
langchain-core
pydantic
langchain-mongodb
genai
langchain-openai
langchain==0.3.27
google-genai==1.43.0
langchain_google_genai==2.0.10
langchain-community==0.3.31